+++ src/functionaltests/java/com/ericsson/ei/subscriptions/trigger/SubscriptionTriggerSteps.java
@@ -1,6 +1,6 @@
 package com.ericsson.ei.subscriptions.trigger;
 
-import com.ericsson.ei.utils.FunctionalBase;
+import com.ericsson.ei.utils.FunctionalTestBase;
 
 import org.junit.Ignore;
 import org.slf4j.Logger;
@@ -14,7 +14,7 @@ import cucumber.api.java.en.When;
 // @TestExecutionListeners(listeners = {
 // DependencyInjectionTestExecutionListener.class,
 // SubscriptionTriggerSteps.class })
-public class SubscriptionTriggerSteps extends FunctionalBase {
+public class SubscriptionTriggerSteps extends FunctionalTestBase {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(SubscriptionTriggerSteps.class);
