--- src/functionaltests/java/com/ericsson/ei/utils/FunctionalBase.java
@@ -1,126 +1,0 @@
-/*
-   Copyright 2017 Ericsson AB.
-   For a full list of individual contributors, please see the commit history.
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-       http://www.apache.org/licenses/LICENSE-2.0
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
-*/
-package com.ericsson.ei.utils;
-
-import com.ericsson.ei.App;
-import com.ericsson.ei.handlers.ObjectHandler;
-import com.ericsson.ei.mongodbhandler.MongoDBHandler;
-import com.ericsson.ei.rmqhandler.RmqHandler;
-import com.ericsson.ei.rules.RulesHandler;
-import com.ericsson.ei.waitlist.WaitListStorageHandler;
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-import java.util.HashMap;
-
-import javax.annotation.PostConstruct;
-
-import org.junit.Ignore;
-import org.junit.runner.RunWith;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.test.context.SpringBootContextLoader;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.TestContext;
-import org.springframework.test.context.TestExecutionListeners;
-import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.test.context.support.AbstractTestExecutionListener;
-import org.springframework.test.context.support.DependencyInjectionTestExecutionListener;
-
-/**
- * @author evasiba
- *
- */
-@Ignore
-@RunWith(SpringRunner.class)
-@SpringBootTest(classes = App.class)
-@ContextConfiguration(classes = App.class, loader = SpringBootContextLoader.class, initializers = TestContextInitializer.class)
-@TestExecutionListeners(listeners = { DependencyInjectionTestExecutionListener.class, FunctionalBase.class })
-public class FunctionalBase extends AbstractTestExecutionListener {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(FunctionalBase.class);
-
-    // @Autowired
-    // private RmqHandler rmqHandler;
-    //
-    // @Autowired
-    // public ObjectHandler objectHandler;
-    //
-    // @Autowired
-    // private RulesHandler rulesHandler;
-    //
-    @Autowired
-    private MongoDBHandler mongoDBHandler;
-    //
-    // @Autowired
-    // private WaitListStorageHandler waitlist;
-
-    // @Autowired
-    // AMQPBrokerManager amqpBroker;
-
-    // @Value("${spring.data.mongodb.database}")
-    // private String database;
-    //
-    // @Value("${event_object_map.collection.name}")
-    // private String event_map;
-    //
-    // private static ObjectMapper objectMapper = new ObjectMapper();
-    //
-    // private static HashMap<String, TestConfigs> configsMap = new
-    // HashMap<String, TestConfigs>();
-    //
-    // @Override
-    // public void beforeTestClass(TestContext testContext) throws Exception {
-    // // System.setProperty("flow.test", "true");
-    // // createFlowTestConfigs();
-    // // getFlowTestConfigs().init();
-    // }
-    //
-    // @PostConstruct
-    // public void init() throws Exception {
-    // // mongoDBHandler.setMongoClient(getFlowTestConfigs().getMongoClient());
-    // // waitlist.setMongoDbHandler(mongoDBHandler);
-    // int i = 0;
-    //
-    // }
-
-    public int getMongoDbPort() {
-        return mongoDBHandler.getPort();
-    }
-
-    // protected TestConfigs getFlowTestConfigs() {
-    // return configsMap.get(getClasName());
-    // }
-    //
-    // private void createFlowTestConfigs() {
-    // TestConfigs newConfigs = new TestConfigs();
-    // String className = getClasName();
-    // configsMap.put(className, newConfigs);
-    // }
-    //
-    // private String getClasName() {
-    // return this.getClass().getName();
-    // }
-    //
-    // private void cleanFlowTestConfigs() {
-    // configsMap.remove(getClasName());
-    // }
-    //
-    // @Override
-    // public void afterTestClass(TestContext testContext) throws Exception {
-    // // cleanFlowTestConfigs();
-    // }
-}
