+++ src/functionaltests/java/com/ericsson/ei/utils/FunctionalTestBase.java
@@ -1,0 +1,51 @@
+/*
+   Copyright 2017 Ericsson AB.
+   For a full list of individual contributors, please see the commit history.
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+       http://www.apache.org/licenses/LICENSE-2.0
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+package com.ericsson.ei.utils;
+
+import com.ericsson.ei.App;
+import com.ericsson.ei.mongodbhandler.MongoDBHandler;
+
+import org.junit.Ignore;
+import org.junit.runner.RunWith;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootContextLoader;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.TestExecutionListeners;
+import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.support.AbstractTestExecutionListener;
+import org.springframework.test.context.support.DependencyInjectionTestExecutionListener;
+
+/**
+ * @author evasiba
+ *
+ */
+@Ignore
+@RunWith(SpringRunner.class)
+@SpringBootTest(classes = App.class)
+@ContextConfiguration(classes = App.class, loader = SpringBootContextLoader.class, initializers = TestContextInitializer.class)
+@TestExecutionListeners(listeners = { DependencyInjectionTestExecutionListener.class, FunctionalTestBase.class })
+public class FunctionalTestBase extends AbstractTestExecutionListener {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(FunctionalTestBase.class);
+
+    @Autowired
+    private MongoDBHandler mongoDBHandler;
+
+    public int getMongoDbPort() {
+        return mongoDBHandler.getPort();
+    }
+}
